tests/ese: crated the first test
