+++ cypress.json
@@ -1,4 +1,12 @@
 {
-  "baseUrl": "http://localhost:4001",
-  "watchForFileChanges": false
+	"watchForFileChanges": false,
+	"baseUrl": "http://localhost:4001",
+	"pageLoadTimeout": 10000,
+	"integrationFolder": "tests/e2e/integration",
+	"supportFile": "tests/e2e/support/index.js",
+	"pluginsFile": "tests/e2e/plugins/index.js",
+	"fixturesFolder": "tests/e2e/fixtures",
+	"fileServerFolder": "log/application-data",
+	"screenshotsFolder": "log/screenshots",
+	"videosFolder": "log/videos"
 }
